server/requirements.txt
fastapi
uvicorn
boto3
mypy-boto3-s3
pandas
python-dotenv
-e .